make api thread safe and validate url parameters

+++ lemon/trading/orders.py
@@ -86,6 +86,9 @@ class Orders:
     def activate(
         self, order_id: str, pin: Optional[str] = None
     ) -> ActivateOrderResponse:
+        if len(order_id) == 0:
+            raise ValueError("order_id is empty string")
+
         resp = self._client.post(
             f"orders/{order_id}/activate",
             json={"pin": pin},
@@ -93,9 +96,15 @@ class Orders:
         return ActivateOrderResponse._from_data(resp.json())
 
     def get_order(self, order_id: str) -> GetOrderResponse:
+        if len(order_id) == 0:
+            raise ValueError("order_id is empty string")
+
         resp = self._client.get(f"orders/{order_id}")
         return GetOrderResponse._from_data(resp.json())
 
     def delete(self, order_id: str) -> DeleteOrderResponse:
+        if len(order_id) == 0:
+            raise ValueError("order_id is empty string")
+
         resp = self._client.delete(f"orders/{order_id}")
         return DeleteOrderResponse._from_data(resp.json())
